pgw#1362 wave 3b: the procsplit cluster folds 5 -> 2, and its shared rig moves to harness/
  5 modules / 2,863 lines  ->  2 modules + 1 shared rig / 2,409 lines
  86 tests in, 85 out

    test_procsplit.py      the split boundary, and the authorization it carries
    test_pod_isolation.py  what the child may touch on the host: uid, OOM rank, host RAM
    harness/split.py       the shared pgw#763 rig (NEW — see below)

A TEST MODULE WAS DOUBLING AS A FIXTURE LIBRARY, and the fold made that impossible to keep. Four
other modules did `from test_procsplit_pgw763 import SplitHarness, captured_dials, ...` — a file
name that stops existing the moment the module is folded. `SplitHarness`, `CHILD_MAIN`, `_payload`,
`captured_dials`, `captured_reports`, `isolated_postmortem` and their dependency closure now live
in `tests/harness/split.py`, where the other shared rigs already live, and test_procsplit,
test_pod_isolation, test_child_faults and test_receipts_trust all import from there.

`TESTS_DIR` had to become `parent.parent`: the hoist moved it one directory deeper, and the
un-fixed version spawned every child from `tests/harness/harness/...`. The suite caught it as a
boot-crash loop, which is the good failure — a path fixture that silently no-ops would not have.

THE VALUE PASS FOUND ALMOST NOTHING, AND THAT IS THE FINDING. Probing `actions.authorize` and
`capability.decide` produced guard-groups of 3 and 8 tests, but each red set spanned DISTINCT
shapes rather than variants. Exactly ONE deletion survived probing: the "unlisted path" row of
`test_action_table_refuses`, whose subject reddens
`test_delta1_parent_refuses_a_hub_call_the_allowlist_does_not_name` — the same property through a
real split child. The "wrong method" row beside it is uniquely guarded (breaking method matching
alone reddens that row and nothing else), so it stays.

THE JOURNEY HYPOTHESIS IS REFUTED FOR THIS CLUSTER. The plan was to replace 23 narrow security rigs
with one broad journey. The probes say those 23 are not variants of each other, so a journey would
have to assert every refusal individually anyway: it would save FIXTURE SETUP, not properties.
Recorded so no later wave re-proposes it from the file count alone.

A COARSE MUTATION FUSES DISTINCT GUARDS — a refinement of 4.34a paid for here. Breaking path AND
method matching together reddened both `test_action_table_refuses` rows and made them look like one
property; narrowing to each separately showed they are two. The probe must be as narrow as the
property it is testing for.

AND A MIS-AIMED PROBE CAME BACK GREEN FOR THE THIRD TIME THIS SESSION: disabling the privilege drop
via a regex that did not apply produced 72 passed. Re-aimed at `plan_drop`'s real early return, 1
RED. A green probe is a claim about the mutation as much as about the tests.

FOLD PROBES: the delta-1 credential strip removed -> 2 RED, one in EACH merged module (the
in-process and containerised halves of the same property). `plan_drop` returning None -> 1 RED.

RATCHETS BOTH TURN THE RIGHT WAY: mypy `ignore_errors` 235 -> 231 with none added; `mypy src tests`
green over 910 files. The pgw#1362 intake baseline shrinks 375 -> 370 — and that gate's FIRST REAL
USE was catching this very fold, refusing the branch until the five folded names left the list.

Counts, per the ruling's format:
  deleted-as-vacuous 0 · deleted-as-low-value 1 · folded-into-module 85 · kept-focused 0
