+++ pyproject.toml
@@ -503,7 +503,6 @@ module = [
   "tests.test_hf_fp8_blockwise_th1803",
   "tests.test_host_canary",
   "tests.test_host_isa_pgw754",
-  "tests.test_host_move_guard_pgw763",
   "tests.test_host_ram_probe_pgw752",
   "tests.test_hot_gpu_pgw636",
   "tests.test_hot_swap_pgw622",
@@ -559,12 +558,9 @@ module = [
   "tests.test_pad32_codegen_pgw1079",
   "tests.test_partial_shape_coverage_pgw844",
   "tests.test_per_stage_staging_pgw1026",
-  "tests.test_pod_privilege_isolation_pgw858",
   "tests.test_pool_sizing_pgw842",
   "tests.test_post_connect_resolution_gate_pgw891",
   "tests.test_precision_resolution_apply",
-  "tests.test_procsplit_pgw763",
-  "tests.test_procsplit_security_pgw763",
   "tests.test_producer_no_shards_th1362",
   "tests.test_provider_routing",
   "tests.test_publish_proof_gate_pgw1191",
